autoclose plugin
autoclose enables textmate/eclipse-style support for parens,
quotes, braces, etc where an automatic closing character is
added to the text when creating the opening character. If you
type the closing quote manually, it will simply overwrite the
quote as you would expect.

credit goes to @julienXX for adding this to his fork