Perform <C-[Ff]> check even if ack isn't installed
Janus sets <C-F> to perform an ack search on non-Mac platforms, but only
if we can map <C-F> and <C-f> to different commands.  The way this was
checked would only work if ack was present, though.  (If ack wasn't
present, <C-f> is mapped to an error message, and not to ":Ack ".)  We
now set <C-F> to a dummy value, perform the <C-F> / <C-f> check, and
then set <C-F>, <C-f>, and <leader>f as appropriate.